Update README.md
Welcome to my portfolio! This repository showcases my journey as a Data Analyst and aspiring Data Scientist. Here, you'll find projects that reflect real-world applications in data analysis, visualization, and engineering. While the data has been anonymized to maintain security and confidentiality, the methodologies and workflows are designed to be transparent, reusable, and practical for similar challenges in any industry.